// packages/sdk-router/src/typechain/FastBridgeV2.ts
export declare namespace IFastBridgeV2 {
  export type BridgeParamsV2Struct = {
    quoteRelayer: string
    quoteExclusivitySeconds: BigNumberish
    quoteId: BytesLike
    zapNative: BigNumberish
    zapData: BytesLike
  }

  export type BridgeParamsV2StructOutput = [
    string,
    BigNumber,
    string,
    BigNumber,
    string
  ] & {
    quoteRelayer: string
    quoteExclusivitySeconds: BigNumber
    quoteId: string
    zapNative: BigNumber
    zapData: string
  }

  export type BridgeTransactionV2Struct = {
    originChainId: BigNumberish
    destChainId: BigNumberish
    originSender: string
    destRecipient: string
    originToken: string
    destToken: string
    originAmount: BigNumberish
    destAmount: BigNumberish
    originFeeAmount: BigNumberish
    deadline: BigNumberish
    nonce: BigNumberish
    exclusivityRelayer: string
    exclusivityEndTime: BigNumberish
    zapNative: BigNumberish
    zapData: BytesLike
  }

  export type BridgeTransactionV2StructOutput = [
    number,
    number,
    string,
    string,
    string,
    string,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    string,
    BigNumber,
    BigNumber,
    string
  ] & {
    originChainId: number
    destChainId: number
    originSender: string
    destRecipient: string
    originToken: string
    destToken: string
    originAmount: BigNumber
    destAmount: BigNumber
    originFeeAmount: BigNumber
    deadline: BigNumber
    nonce: BigNumber
    exclusivityRelayer: string
    exclusivityEndTime: BigNumber
    zapNative: BigNumber
    zapData: string
  }
}